Se trabajo en la conexion usuario, carrito, pagos y ordenes la conexion no marca errores pero no se refleja en postman ni en la base datos por lo que tengo que trabajar para encontrar el error de la conexion, se hizo cambios en los db/migration para que tengan correlacion entre ellos, los metodos por si mismos funcionan pero las conexiones dan fallo.

+++ carrito/src/main/java/cl/duoc/carrito/dto/CarritoResponse.java
@@ -1,5 +1,7 @@
 package cl.duoc.carrito.dto;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -12,7 +14,7 @@ import lombok.NoArgsConstructor;
 public class CarritoResponse {
     private Long id;
     private int cantidad;
-    private int montoTotal;
+    private BigDecimal montoTotal;
     private UsuarioResponse usuario; 
     private ProductoResponse producto;
 }
